requirements.txt
Pillow==10.3.0
PyPDF2==3.0.1
Pyphen==0.15.0
pytesseract==0.3.10
